requirements.txt
asgiref
Django
sqlparse
typing_extensions
libsass
django-extensions
pyOpenSSL
python-jose
PyJWT
requests
xmltodict
design-django-theme
Pillow
djangorestframework
apscheduler
markdown-it-py
cryptography
openpyxl
pyzipper
html2text
django-sass-processor
django-compressor
gunicorn
python-decouple
whitenoise
pytz
